fix: 修复代码审查中发现的5个小问题
1. search.cpp: 修复 j<6 依赖检测的移位比较错误（假阳性），改用
   正确的 mask0/mask1 配对位比较
2. affine.cpp: 移除 m>25 无一致性校验的 XOR 路径，替换为使用
   unordered_map 的一致性校验路径（m>27 时避免 2^m 内存）
3. affine.cpp: 修复 save_g_tt=false 时 union_T = total_T 的近似值，
   改为始终计算精确 union_T
4. main_gate_builder.cpp: 移除不再被调用的旧版 write_affine/
   write_poly/write_stats 死代码
5. main_opt2.cpp: 修复 mono_mask 的 int 截断溢出风险，改为 int64_t

+++ cpp/src/main_opt2.cpp
@@ -501,7 +501,7 @@ int main(int argc, char** argv) {
             while (word) {
                 int bit_idx = __builtin_ctzll(word);
                 word &= word - 1;
-                int mono_mask = (int)((w << 6) | bit_idx);
+                int64_t mono_mask = (int64_t)((w << 6) | bit_idx);
 
                 std::vector<int> svars;
                 int rem = mono_mask;
